feat(httpclient): add ability to form encode parameters (#543)
* feat(httpclient): add ability to form encode parameters

== Motivation ==

Support form encoding of query params

== Details ==

This addresses https://github.com/ClickHouse/clickhouse-connect/issues/342

Since https://github.com/ClickHouse/ClickHouse/issues/8842, query parameters
can now be sent to ClickHouse when using HTTP via the body as form encoded
paramters. This change adds the ability to enable this feature at a client level.
Making this available per query (i.e. via the method parameters of the query
methods) could be added as a future improvement, but this would require
a larger refactor. So for now we are making it a client wide setting AND
only available for querying (and not inserting).

== Testing ==

Added and ran integration tests

* update test parameter naming

* move query to body when using form_encode_query_params and no external_data is present

* add external_data + form_encode_query_params integration test

* add unit tests, code reoganization for easier understanding

* address feedback

* address external_data support in schema probe

---------

Signed-off-by: Joe Spadola <joe.spadola@clickhouse.com>
Co-authored-by: sanjams2 <sanjams2@users.noreply.github.com>
Co-authored-by: Joe Spadola <joe.spadola@clickhouse.com>